+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 [Commits](https://github.com/thebigmunch/audio-metadata/compare/0.6.0...master)
 
+### Added
+
+* ID3 version as attribute on ``ID3v2Frames``.
+
 ### Changed
 
 * Set ``FIELD_MAP`` on instance rather than class in ``ID3v2Frames.load``.